backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
PyMySQL==1.1.0
redis==5.0.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
celery==5.3.2
plotly==5.17.0
statsmodels==0.14.0
werkzeug==2.3.7